+++ tests/tinytest/test_across_lazy.R
@@ -281,4 +281,30 @@ expect_colnames(
   setdiff(colnames(mtcars), "gear")
 )
 
+
+# Need to specify .cols (either named or unnamed)
+
+expect_error_lazy(
+  mutate(test, across(.fns = mean)),
+  "You must supply the argument `.cols`"
+)
+
+# test .by
+
+test3 <- mtcars |>
+  head(n = 5) |>
+  as_polars() |>
+  mutate(across(everything(), .fns = mean), .by = "cyl") |>
+  distinct()
+
+expect_equal_lazy(
+  test3 |> pull(cyl),
+  c(6, 4, 8)
+)
+
+expect_equal_lazy(
+  test3 |> pull(hp),
+  c(110, 93, 175)
+)
+
 Sys.setenv('TIDYPOLARS_TEST' = FALSE)